added save on game over and  game complete

+++ Assets/Scripts/Data/PlayerData.cs
@@ -6,7 +6,7 @@ public class PlayerData
     public string name;
     public float currentCoins;
     public int[] levelStars;      // stars per level
-    public int[] levelPoints;     // points per level
+    public float[] levelPoints;     // points per level
     public int unlockedLevel = 1;
     public float musicVolume;
     public float sfxVolume;
